TgBot/requirements.txt
python-telegram-bot==20.6
fastapi
uvicorn
psycopg2-binary
python-dotenv